cs40l25: Add primitive delays for R11
R11 sometimes hits polling timeout before playing the whole queue. This happens because the actual time it takes to play an effect is greater than the expected timeline, and on larger effect queues this means some effects will not be played, which impacts features such as Vibration Watch, that tells the time based on the number of vibrations. The whole investigation and delay values are detailed in b/288487031.

Because there are other devices that use cs40l25, the delays are added only if the ro.vendor.vibrator.hal.cs40L25.primitive_delays.enabled property is set to true.

Bug:321107928
Test: Check in logs that polling timeout doesn't affect effect playing.
Change-Id: I4d0430da1e0ef90852b15f8e2fbf0e34f5f61f28
Signed-off-by: Eduard Olteanu <eduardolteanu@google.com>