thermal: support nested thermal configs
Device thermal configs have grown over the years to support multiple
features. We have multiple configs catering to hw revisions and device
operating modes. A lot of configuration is common across the configs
and it is repeated in each config.

This feature supports nested configs. All the common configuration can
be moved to a base config and this config can be included in all the
parent thermal configs.

Added support for 'isDisabled' flag to be able to disable sensors or
cooling devices in the base config at top level config

Bug: 306185660
Test: Build and boot on device

Change-Id: Iedb36878aac837dbd2a2c4c1c4658b51d499833c
Signed-off-by: Sayanna Chandula <sayanna@google.com>